ar/python-core-framework/pyproject.toml
[tool.poetry]
name = "core"
version = "2022.1.11"
description = "Python Core Kit"
authors = ["Alejandro M. BERNARDIS <alejandro.bernardis@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aetcd3 = "^0.1.0-alpha.7"
aiodns = "^3.0.0"
aiohttp = "^3.8.1"
aiohttp_cors = "^0.7.0"
aiojobs = "^1.0.0"
aiomysql = "^0.0.22"
aioredis = "^2.0.0"
async_lru = "^1.0.2"
brotlipy = "^0.7.0"
cchardet = "^2.1.7"
cryptography = "^36.0.1"
docopt = "^0.6.2"
motor = "^2.5.1"
orjson = "^3.6.5"
pydantic = "^1.8.2"
pytest-aiohttp = "^0.3.0"
pytest = "^6.2.5"
python-dotenv = "^0.19.2"
pytz = "^2021.3"
uvloop = "^0.16.0"
yarl = "^1.7.2"
six = "^1.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
